requirements.txt
python-dotenv
langchain
langchain-community
langchain-huggingface
torch
faiss-cpu
streamlit
requests
tqdm
pypdf
beautifulsoup4